web_login/zhihu/requirements.txt
requests==2.19.1
matplotlib==3.1.1
PyExecJS==1.5.1
Pillow==6.1.0
